Map data type xsd:decimal to python Decimal

+++ tests/fixtures/defxmlschema/chapter12/example1232.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from dataclasses import dataclass, field
 from typing import Optional
 
@@ -17,7 +18,7 @@ class ProductType:
             required=True
         )
     )
-    version: Optional[float] = field(
+    version: Optional[Decimal] = field(
         default=None,
         metadata=dict(
             name="version",
