+++ tests/fixtures/defxmlschema/chapter15/example15171.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from dataclasses import dataclass, field
 from lxml.etree import QName
 from typing import Dict, Optional
@@ -20,7 +21,7 @@ class PurchaseOrderType:
             required=True
         )
     )
-    version: Optional[float] = field(
+    version: Optional[Decimal] = field(
         default=None,
         metadata=dict(
             name="version",
